Add feedback after test connection
