+++ airbyte-webapp/src/components/ServiceForm/components/BottomBlock.tsx
@@ -3,8 +3,9 @@ import styled from "styled-components";
 import { FormattedMessage } from "react-intl";
 
 import Button from "components/Button";
-import Spinner from "components/Spinner";
-import StatusIcon from "components/StatusIcon";
+import TestingConnectionSpinner from "./TestingConnectionSpinner";
+import TestingConnectionSuccess from "./TestingConnectionSuccess";
+import TestingConnectionError from "./TestingConnectionError";
 
 type IProps = {
   formType: "source" | "destination" | "connection";
@@ -22,47 +23,6 @@ const ButtonContainer = styled.div`
   justify-content: space-between;
 `;
 
-const Success = styled(StatusIcon)`
-  width: 26px;
-  min-width: 26px;
-  height: 26px;
-  padding-top: 5px;
-  font-size: 17px;
-`;
-
-const LoadingContainer = styled(ButtonContainer)`
-  font-weight: 600;
-  font-size: 14px;
-  line-height: 17px;
-  color: ${({ theme }) => theme.darkPrimaryColor};
-  justify-content: center;
-`;
-
-const Loader = styled.div`
-  margin-right: 10px;
-`;
-
-const Error = styled(Success)`
-  padding-top: 4px;
-  padding-left: 1px;
-`;
-
-const ErrorBlock = styled.div`
-  display: flex;
-  justify-content: right;
-  align-items: center;
-  font-weight: 600;
-  font-size: 12px;
-  line-height: 18px;
-  color: ${({ theme }) => theme.darkPrimaryColor};
-`;
-
-const ErrorText = styled.div`
-  font-weight: normal;
-  color: ${({ theme }) => theme.dangerColor};
-  max-width: 400px;
-`;
-
 const BottomBlock: React.FC<IProps> = ({
   isSubmitting,
   formType,
@@ -72,35 +32,17 @@ const BottomBlock: React.FC<IProps> = ({
   additionBottomControls,
 }) => {
   if (hasSuccess) {
-    return (
-      <LoadingContainer>
-        <Success success />
-        <FormattedMessage id="form.successTests" />
-      </LoadingContainer>
-    );
+    return <TestingConnectionSuccess />;
   }
 
   if (isSubmitting) {
-    return (
-      <LoadingContainer>
-        <Loader>
-          <Spinner />
-        </Loader>
-        <FormattedMessage id="form.testingConnection" />
-      </LoadingContainer>
-    );
+    return <TestingConnectionSpinner />;
   }
 
   return (
     <ButtonContainer>
       {errorMessage ? (
-        <ErrorBlock>
-          <Error />
-          <div>
-            <FormattedMessage id="form.failedTests" />
-            <ErrorText>{errorMessage}</ErrorText>
-          </div>
-        </ErrorBlock>
+        <TestingConnectionError errorMessage={errorMessage} />
       ) : (
         <div />
       )}
