+++ airbyte-webapp/src/components/ServiceForm/components/EditControls.tsx
@@ -3,8 +3,10 @@ import styled from "styled-components";
 import { FormattedMessage } from "react-intl";
 
 import Button from "components/Button";
-// import Spinner from "components/Spinner";
 import { useServiceForm } from "../serviceFormContext";
+import TestingConnectionSpinner from "./TestingConnectionSpinner";
+import TestingConnectionSuccess from "./TestingConnectionSuccess";
+import TestingConnectionError from "./TestingConnectionError";
 
 type IProps = {
   isSubmitting: boolean;
@@ -29,26 +31,6 @@ const ButtonContainer = styled.span`
   margin-left: 10px;
 `;
 
-// const Success = styled(ButtonContainer)`
-//   color: ${({ theme }) => theme.successColor};
-//   font-size: 14px;
-//   line-height: 17px;
-// `;
-//
-// const Error = styled(Success)`
-//   color: ${({ theme }) => theme.dangerColor};
-//   max-width: 68%;
-//   display: inline-block;
-//   vertical-align: middle;
-// `;
-//
-// const SpinnerContainer = styled.div`
-//   margin: -13px 0 0 10px;
-//   display: inline-block;
-//   position: relative;
-//   top: 10px;
-// `;
-
 const EditControls: React.FC<IProps> = ({
   isSubmitting,
   isValid,
@@ -56,60 +38,58 @@ const EditControls: React.FC<IProps> = ({
   resetForm,
   formType,
   onRetest,
-  // successMessage,
-  // errorMessage,
+  successMessage,
+  errorMessage,
 }) => {
   const { unfinishedSecrets } = useServiceForm();
 
-  // const showStatusMessage = () => {
-  //   if (isSubmitting) {
-  //     return (
-  //       <SpinnerContainer>
-  //         <Spinner small />
-  //       </SpinnerContainer>
-  //     );
-  //   }
-  //   if (errorMessage) {
-  //     return <Error>{errorMessage}</Error>;
-  //   }
-  //   if (successMessage && !dirty) {
-  //     return <Success data-id="success-result">{successMessage}</Success>;
-  //   }
-  //   return null;
-  // };
+  if (isSubmitting) {
+    return <TestingConnectionSpinner />;
+  }
+  const showStatusMessage = () => {
+    if (errorMessage) {
+      return <TestingConnectionError errorMessage={errorMessage} />;
+    }
+    if (successMessage) {
+      return <TestingConnectionSuccess />;
+    }
+    return null;
+  };
 
   return (
-    <Controls>
-      <div>
-        <Button
-          type="submit"
-          disabled={
-            isSubmitting ||
-            !isValid ||
-            !dirty ||
-            Object.keys(unfinishedSecrets).length > 0
-          }
-        >
-          <FormattedMessage id="form.saveChangesAndTest" />
-        </Button>
-        <ButtonContainer>
+    <>
+      {showStatusMessage()}
+      <Controls>
+        <div>
           <Button
-            type="button"
-            secondary
-            disabled={isSubmitting || !isValid || !dirty}
-            onClick={resetForm}
+            type="submit"
+            disabled={
+              isSubmitting ||
+              !isValid ||
+              !dirty ||
+              Object.keys(unfinishedSecrets).length > 0
+            }
           >
-            <FormattedMessage id="form.cancel" />
+            <FormattedMessage id="form.saveChangesAndTest" />
           </Button>
-        </ButtonContainer>
-        {/*{showStatusMessage()}*/}
-      </div>
-      {onRetest && (
-        <Button type="button" onClick={onRetest}>
-          <FormattedMessage id={`form.${formType}Retest`} />
-        </Button>
-      )}
-    </Controls>
+          <ButtonContainer>
+            <Button
+              type="button"
+              secondary
+              disabled={isSubmitting || !isValid || !dirty}
+              onClick={resetForm}
+            >
+              <FormattedMessage id="form.cancel" />
+            </Button>
+          </ButtonContainer>
+        </div>
+        {onRetest && (
+          <Button type="button" onClick={onRetest} disabled={!isValid}>
+            <FormattedMessage id={`form.${formType}Retest`} />
+          </Button>
+        )}
+      </Controls>
+    </>
   );
 };
